Added copyCommandToClipboard buttons in scene.debugLayer Inspector
