+++ packages/dev/inspector/src/components/actionTabs/actionTabs.scss
@@ -676,11 +676,11 @@ $line-padding-left: 2px;
                     }
 
                     .radioContainer {
+                        grid-column: 2;
                         display: flex;
                         align-items: center;
 
                         .radio {
-                            grid-column: 2;
                             display: none;
 
                             &:checked + label:before {
@@ -724,6 +724,17 @@ $line-padding-left: 2px;
                             }
                         }
                     }
+
+                    .copy {
+                        grid-column: 3;
+                        display: grid;
+                        align-items: center;
+                        justify-items: center;
+                        cursor: pointer;
+                        img {
+                            height: 100%;
+                        }
+                    }
                 }
 
                 .vector3Line {
@@ -759,6 +770,19 @@ $line-padding-left: 2px;
                             justify-items: center;
                             cursor: pointer;
                         }
+
+                        .copy {
+                            grid-column: 4;
+                            display: grid;
+                            align-items: center;
+                            justify-items: center;
+                            cursor: pointer;
+                            img {
+                                height: 100%;
+                            }
+                        }
+
+
                     }
 
                     .secondLine {
@@ -794,7 +818,7 @@ $line-padding-left: 2px;
                     padding-left: $line-padding-left;
                     height: 30px;
                     display: grid;
-                    grid-template-columns: 1fr auto;
+                    grid-template-columns: 1fr auto 20px 10px;
 
                     .label {
                         grid-column: 1;
@@ -868,6 +892,18 @@ $line-padding-left: 2px;
 
                         .icon {
                             display: none;
+                        }
+
+                    }
+
+                    .copy {
+                        grid-column: 3;
+                        display: grid;
+                        align-items: center;
+                        justify-items: center;
+                        cursor: pointer;
+                        img {
+                            height: 100%;
                         }
                     }
                 }
@@ -1072,6 +1108,17 @@ $line-padding-left: 2px;
                             cursor: pointer;
                         }
                     }
+
+                    .copy {
+                        grid-column: 4;
+                        display: grid;
+                        align-items: center;
+                        justify-items: center;
+                        cursor: pointer;
+                        img {
+                            height: 100%;
+                        }
+                    }
                 }
 
                 .color3Line {
@@ -1081,7 +1128,7 @@ $line-padding-left: 2px;
                     .firstLine {
                         height: 30px;
                         display: grid;
-                        grid-template-columns: 1fr auto 20px 20px;
+                        grid-template-columns: 1fr auto 20px;
 
                         .label {
                             grid-column: 1;
@@ -1115,8 +1162,16 @@ $line-padding-left: 2px;
                             }
                         }
 
-                        .copy {
+                        .expand {
                             grid-column: 3;
+                            display: grid;
+                            align-items: center;
+                            justify-items: center;
+                            cursor: pointer;
+                        }
+
+                        .copy {
+                            grid-column: 4;
                             display: grid;
                             align-items: center;
                             justify-items: center;
@@ -1126,13 +1181,6 @@ $line-padding-left: 2px;
                             }
                         }
 
-                        .expand {
-                            grid-column: 4;
-                            display: grid;
-                            align-items: center;
-                            justify-items: center;
-                            cursor: pointer;
-                        }
                     }
 
                     .secondLine {
@@ -1168,7 +1216,7 @@ $line-padding-left: 2px;
                     padding-left: $line-padding-left;
                     height: 30px;
                     display: grid;
-                    grid-template-columns: 1fr auto;
+                    grid-template-columns: 1fr auto 20px 10px;
 
                     .label {
                         grid-column: 1;
@@ -1185,6 +1233,17 @@ $line-padding-left: 2px;
 
                         select {
                             width: 115px;
+                        }
+                    }
+
+                    .copy {
+                        grid-column: 3;
+                        display: grid;
+                        align-items: center;
+                        justify-items: center;
+                        cursor: pointer;
+                        img {
+                            height: 100%;
                         }
                     }
                 }
