+++ packages/dev/sharedUiComponents/src/lines/sliderLineComponent.tsx
@@ -1,9 +1,12 @@
 import * as React from "react";
 import type { Observable } from "core/Misc/observable";
 import type { PropertyChangedEvent } from "../propertyChangedEvent";
+import { copyCommandToClipboard } from "../copyCommandToClipboard";
 import { Tools } from "core/Misc/tools";
 import { FloatLineComponent } from "./floatLineComponent";
 import type { LockObject } from "../tabs/propertyGrids/lockObject";
+
+import copyIcon from "./copy.svg";
 
 interface ISliderLineComponentProps {
     label: string;
@@ -120,6 +123,16 @@ export class SliderLineComponent extends React.Component<ISliderLineComponentPro
         return value;
     }
 
+    onCopyClick(){
+        if(this.props && this.props.target){
+            let targetName = this.props.target.constructor.name;
+            let targetProperty = this.props.propertyName;
+            let value = this.props.target[this.props.propertyName!];
+            let strCommand = targetName + "." + targetProperty + " = " + value + ";";
+            copyCommandToClipboard(strCommand);
+        }
+    }
+
     override render() {
         return (
             <div className="sliderLine">
@@ -162,7 +175,11 @@ export class SliderLineComponent extends React.Component<ISliderLineComponentPro
                         onChange={(evt) => this.onChange(evt.target.value)}
                     />
                 </div>
+                <div className="copy hoverIcon" onClick={() => this.onCopyClick()} title="Copy to clipboard">
+                    <img src={copyIcon} alt="Copy" />
+                </div>
             </div>
+
         );
     }
 }
